:art: moves valitdation and validation components to separate files
