+++ src/forms/index.js
@@ -1,4 +1,5 @@
-export {default as EditDocumentForm} from './EditDocumentForm';
+export {default as EditDocumentForm} from './EditDocument/EditDocumentForm';
+export {default as EditStudyForm} from './EditStudyForm';
 export {default as NewProjectForm} from './NewProjectForm';
 export {default as LinkProjectForm} from './LinkProjectForm';
 export {default as EditProjectForm} from './EditProjectForm';
